+++ Delfi.Glo.Repository/IWellDetailsInfoService.cs
@@ -1,9 +1,13 @@
 ﻿
+using Delfi.Glo.Entities.Dto;
+
 namespace Delfi.Glo.Repository
 {
   
     public interface IWellDetailsInfoService<T> where T : class
-    { 
-        Task<T> GetSwimLaneDetailsByDate(int WellId, DateTime StartDate, DateTime EndDate);
+    {
+        Task<T> GetWellDetailsInfoById(int WellId);
+        Task<SwimLaneGraphDetails> GetSwimLaneDetailsByDate(int WellId, DateTime StartDate, DateTime EndDate);
+
     }
 }
