+++ Delfi.Glo.Repository/IWellService.cs
@@ -1,9 +1,13 @@
-﻿namespace Delfi.Glo.Repository
+﻿using Delfi.Glo.Common.Services;
+using Delfi.Glo.Entities.Dto;
+
+namespace Delfi.Glo.Repository
 {
     public interface IWellService<T> where T : class
     {
-        Task<T> GetWellDetailsInfoById(int WellId);
-
-      //  Task<T> GetSetPointDetailsByTotalUpTime(int WellId, DateTime StartDate, DateTime EndDate);
+        Task<Tuple<bool, IEnumerable<WellDto>, int, int, int, int>> GetWellListByFilter(int page, int pageSize, string? searchString, string? ApprovalStatus, string? ApprovalMode, List<SortExpression> sortExpression);
+        Task<IEnumerable<WellDto>> GetAllAsync();
+        Task<IEnumerable<WellDto>> GetWells();
+        Task<WellDto> GetAsync(int id);
     }
 }
